requirements.txt
selenium~=4.2.0
webdriver-manager~=3.7.0
pytest-html~=3.1.1